Relocate pinch
