+++ src/components/page-viewer/PageViewer.tsx
@@ -2,7 +2,7 @@ import { Dimensions } from '@/types/dimensions.interface'
 import PageScroller from './PageScroller'
 import { useRef, useState } from 'react'
 import ImgWrapper from './ImgWrapper'
-import { usePinchPan } from '@/hooks/pinch-pan/use-pinch-pan'
+import { usePinchPan } from '@/hooks/use-pinch-pan'
 import { useScrollingManager } from './use-scrolling-manager'
 import { usePinchingManager } from './use-pinching-manager'
 import { useScrollLimits } from './use-scroll-limits'
