+++ src/hooks/use-pinch-pan/point-utils.ts
