+++ src/hooks/use-pinch-pan/use-pointer-tracker.ts
